feat: add popup
